web/requirements.txt
Flask==3.0.1
psycopg2-binary
pytest
pytest-cov
allure-pytest
